+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to GOBL will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/). See also the [GOBL versions](https://docs.gobl.org/overview/versions) documentation site for more details.
 
+## [Unreleased]
+
+### Added
+
+- `br-nfse-v1`: added initial Brazil NFS-e addon
+
 ## [v0.203.0]
 
 ### Added